Moved tests to new file
